Combine server/index.js and server/index.js
